mesh navigation
Implemented a few data structures/algorithms to map the hexa mesh
structure. Now quads are no longer needed in the mesh file in order to
be able to do the rendering, they can be extracted from the hexas.
Hopefully in the future this stuff will prove itself useful and handy.